+++ src/rpc/info.rs
@@ -4,7 +4,7 @@ use crate::Id;
 
 use super::Rpc;
 
-/// Information and statistics about this [Dht] node.
+/// Information and statistics about this mainline node.
 #[derive(Debug, Clone)]
 pub struct Info {
     id: Id,
@@ -26,7 +26,7 @@ impl Info {
     }
     /// Returns the best guess for this node's Public addresss.
     ///
-    /// If [Config::public_ip] was set, this is what will be returned
+    /// If [super::Config::public_ip] was set, this is what will be returned
     /// (plus the local port), otherwise it will rely on consensus from
     /// responding nodes voting on our public IP and port.
     pub fn public_address(&self) -> Option<SocketAddrV4> {
@@ -35,7 +35,7 @@ impl Info {
     /// Returns `true` if we can't confirm that [Self::public_address] is publicly addressable.
     ///
     /// If this node is firewalled, it won't switch to server mode if it is in adaptive mode,
-    /// but if [Config::server_mode] was set to true, then whether or not this node is firewalled
+    /// but if [super::Config::server_mode] was set to true, then whether or not this node is firewalled
     /// won't matter.
     pub fn firewalled(&self) -> bool {
         self.firewalled